+++ example/lib/flap/flap_home_page.dart
@@ -1,8 +1,6 @@
 import 'package:bitsdojo_window/bitsdojo_window.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:libadwaita/libadwaita.dart';
-import 'package:window_decorations/window_decorations.dart';
 
 class FlapHomePage extends StatefulWidget {
   const FlapHomePage({Key? key, required this.themeNotifier}) : super(key: key);
@@ -51,7 +49,6 @@ class _FlapHomePageState extends State<FlapHomePage> {
       children: [
         AdwHeaderBar.bitsdojo(
           appWindow: appWindow,
-          windowDecor: windowDecor,
           start: [
             Builder(
               builder: (context) {
@@ -132,7 +129,7 @@ class _FlapHomePageState extends State<FlapHomePage> {
                           title: 'Locked',
                           subtitle: """
 Sidebar visibility doesn't change when fold state changes""",
-                          end: CupertinoSwitch(
+                          end: AdwSwitch(
                             value: locked,
                             onChanged: (val) {
                               locked = val;
@@ -180,18 +177,15 @@ Sidebar visibility doesn't change when fold state changes""",
                           title: 'Modal',
                           subtitle: '''
 Clicking outside the sidebar or pressing Esc will close it when folded''',
-                          end:
-                              CupertinoSwitch(value: true, onChanged: (val) {}),
+                          end: AdwSwitch(value: true, onChanged: (val) {}),
                         ),
                         AdwActionRow(
                           title: 'Swipe to Open',
-                          end:
-                              CupertinoSwitch(value: true, onChanged: (val) {}),
+                          end: AdwSwitch(value: true, onChanged: (val) {}),
                         ),
                         AdwActionRow(
                           title: 'Swipe to Close',
-                          end:
-                              CupertinoSwitch(value: true, onChanged: (val) {}),
+                          end: AdwSwitch(value: true, onChanged: (val) {}),
                         ),
                       ],
                     ),