+++ example/lib/home_page.dart
@@ -9,7 +9,7 @@ import 'package:example/pages/view_switcher_page.dart';
 import 'package:example/pages/welcome.dart';
 import 'package:flutter/material.dart';
 import 'package:libadwaita/libadwaita.dart';
-import 'package:window_decorations/window_decorations.dart';
+import 'package:url_launcher/url_launcher.dart';
 
 class MyHomePage extends StatefulWidget {
   const MyHomePage({Key? key, required this.themeNotifier}) : super(key: key);
@@ -52,11 +52,18 @@ class _MyHomePageState extends State<MyHomePage> {
 
   @override
   Widget build(BuildContext context) {
+    final developers = {
+      'Prateek Sunal': 'prateekmedia',
+      'Malcolm Mielle': 'MalcolmMielle',
+      'sim': 'simrat39',
+      'Jesús Rodríguez': 'jesusrp98',
+      'Polo': 'pablojimpas',
+    };
+
     return Column(
       children: [
         AdwHeaderBar.bitsdojo(
           appWindow: appWindow,
-          windowDecor: windowDecor,
           start: [
             Builder(
               builder: (context) {
@@ -79,14 +86,66 @@ class _MyHomePageState extends State<MyHomePage> {
             AdwPopupMenu(
               body: Column(
                 mainAxisSize: MainAxisSize.min,
+                crossAxisAlignment: CrossAxisAlignment.stretch,
                 children: [
-                  ListTile(
-                    onTap: () {
+                  AdwButton.flat(
+                    onPressed: () {
                       counter.value = 0;
                       Navigator.of(context).pop();
                     },
-                    title: const Text(
+                    padding: AdwButton.defaultButtonPadding.copyWith(
+                      top: 10,
+                      bottom: 10,
+                    ),
+                    child: const Text(
                       'Reset Counter',
+                      style: TextStyle(fontSize: 15),
+                    ),
+                  ),
+                  const Divider(),
+                  AdwButton.flat(
+                    padding: AdwButton.defaultButtonPadding.copyWith(
+                      top: 10,
+                      bottom: 10,
+                    ),
+                    child: const Text(
+                      'Preferences',
+                      style: TextStyle(fontSize: 15),
+                    ),
+                  ),
+                  AdwButton.flat(
+                    padding: AdwButton.defaultButtonPadding.copyWith(
+                      top: 10,
+                      bottom: 10,
+                    ),
+                    onPressed: () => showDialog<Widget>(
+                      context: context,
+                      builder: (ctx) => AdwAboutWindow(
+                        issueTrackerLink:
+                            'https://github.com/gtk-flutter/libadwaita/issues',
+                        appIcon: Image.asset('assets/logo.png'),
+                        credits: [
+                          AdwPreferencesGroup.credits(
+                            title: 'Developers',
+                            children: developers.entries
+                                .map(
+                                  (e) => AdwActionRow(
+                                    title: e.key,
+                                    onActivated: () =>
+                                        launch('https://github.com/${e.value}'),
+                                  ),
+                                )
+                                .toList(),
+                          ),
+                        ],
+                        copyright: 'Copyright 2021-2022 Gtk-Flutter Developers',
+                        license: const Text(
+                          'GNU LGPL-3.0, This program comes with no warranty.',
+                        ),
+                      ),
+                    ),
+                    child: const Text(
+                      'About this Demo',
                       style: TextStyle(fontSize: 15),
                     ),
                   ),