+++ lib/src/utils/colors.dart
@@ -27,20 +27,22 @@ Color borderLight = Colors.black.withOpacity(0.18);
 Color borderDark = const Color(0xFF454545);
 
 extension BorderContext on BuildContext {
-  Color get borderColor =>
-      Theme.of(this).brightness == Brightness.dark ? borderDark : borderLight;
-}
+  bool get _isDark => Theme.of(this).brightness == Brightness.dark;
 
-extension SelectContext on BuildContext {
-  Color get selectColor => Theme.of(this).brightness == Brightness.dark
+  Color get borderColor => _isDark ? borderDark : borderLight;
+  Color get checkboxColor =>
+      _isDark ? const Color(0xFF535353) : const Color(0xFFE0E0E0);
+
+  Color get selectColor => _isDark
       ? Theme.of(this).backgroundColor.lighten()
       : Theme.of(this).backgroundColor.darken();
 
-  Color get hoverMenuColor => Theme.of(this).brightness == Brightness.dark
+  Color get hoverMenuColor => _isDark
       ? Theme.of(this).backgroundColor.lighten(0.05)
       : Theme.of(this).backgroundColor.darken(0.05);
 
-  Color get hoverColor => Theme.of(this).brightness == Brightness.dark
+  Color get hoverColor => _isDark
       ? Theme.of(this).backgroundColor.lighten(0.15)
       : Theme.of(this).backgroundColor.darken(0.005);
+  Color get textColor => _isDark ? Colors.white : Colors.black;
 }