+++ lib/src/widgets/adw/avatar.dart
@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 /// color combinations a `AdwAvatar` widget
 /// can feature. They're extracted directly
 /// from the GTK `libadwaita` implementation.
-enum AdwAvatarColors {
+enum AdwColors {
   blue,
   cyan,
   green,
@@ -23,7 +23,7 @@ enum AdwAvatarColors {
   gray
 }
 
-extension AvatarColorsExtension on AdwAvatarColors? {
+extension AvatarColorsExtension on AdwColors? {
   AdwAvatarColor get getColor => AdwAvatarColorPalette.getColor(this);
 
   Color get foregroundColor => getColor.foregroundColor;
@@ -49,98 +49,98 @@ class AdwAvatarColor {
 /// and background of a `AdwAvatar` widget.
 class AdwAvatarColorPalette {
   static const _palette = {
-    AdwAvatarColors.blue: AdwAvatarColor(
+    AdwColors.blue: AdwAvatarColor(
       foregroundColor: Color(0xFFcfe1f5),
       backgroundGradient: [
         Color(0xFF83b6ec),
         Color(0xFF337fdc),
       ],
     ),
-    AdwAvatarColors.cyan: AdwAvatarColor(
+    AdwColors.cyan: AdwAvatarColor(
       foregroundColor: Color(0xFFcaeaf2),
       backgroundGradient: [
         Color(0xFF7ad9f1),
         Color(0xFF0f9ac8),
       ],
     ),
-    AdwAvatarColors.green: AdwAvatarColor(
+    AdwColors.green: AdwAvatarColor(
       foregroundColor: Color(0xFFcef8d8),
       backgroundGradient: [
         Color(0xFF8de6b1),
         Color(0xFF29ae74),
       ],
     ),
-    AdwAvatarColors.lime: AdwAvatarColor(
+    AdwColors.lime: AdwAvatarColor(
       foregroundColor: Color(0xFFe6f9d7),
       backgroundGradient: [
         Color(0xFFb5e98a),
         Color(0xFF6ab85b),
       ],
     ),
-    AdwAvatarColors.yellow: AdwAvatarColor(
+    AdwColors.yellow: AdwAvatarColor(
       foregroundColor: Color(0xFFf9f4e1),
       backgroundGradient: [
         Color(0xFFf8e359),
         Color(0xFFd29d09),
       ],
     ),
-    AdwAvatarColors.gold: AdwAvatarColor(
+    AdwColors.gold: AdwAvatarColor(
       foregroundColor: Color(0xFFffead1),
       backgroundGradient: [
         Color(0xFFffcb62),
         Color(0xFFd68400),
       ],
     ),
-    AdwAvatarColors.orange: AdwAvatarColor(
+    AdwColors.orange: AdwAvatarColor(
       foregroundColor: Color(0xFFffe5c5),
       backgroundGradient: [
         Color(0xFFffa95a),
         Color(0xFFed5b00),
       ],
     ),
-    AdwAvatarColors.raspberry: AdwAvatarColor(
+    AdwColors.raspberry: AdwAvatarColor(
       foregroundColor: Color(0xFFf8d2ce),
       backgroundGradient: [
         Color(0xFFf78773),
         Color(0xFFe62d42),
       ],
     ),
-    AdwAvatarColors.magenta: AdwAvatarColor(
+    AdwColors.magenta: AdwAvatarColor(
       foregroundColor: Color(0xFFfac7de),
       backgroundGradient: [
         Color(0xFFe973ab),
         Color(0xFFe33b6a),
       ],
     ),
-    AdwAvatarColors.purple: AdwAvatarColor(
+    AdwColors.purple: AdwAvatarColor(
       foregroundColor: Color(0xFFe7c2e8),
       backgroundGradient: [
         Color(0xFFcb78d4),
         Color(0xFF9945b5),
       ],
     ),
-    AdwAvatarColors.violet: AdwAvatarColor(
+    AdwColors.violet: AdwAvatarColor(
       foregroundColor: Color(0xFFd5d2f5),
       backgroundGradient: [
         Color(0xFF9e91e8),
         Color(0xFF7a59ca),
       ],
     ),
-    AdwAvatarColors.beige: AdwAvatarColor(
+    AdwColors.beige: AdwAvatarColor(
       foregroundColor: Color(0xFFf2eade),
       backgroundGradient: [
         Color(0xFFe3cf9c),
         Color(0xFFb08952),
       ],
     ),
-    AdwAvatarColors.brown: AdwAvatarColor(
+    AdwColors.brown: AdwAvatarColor(
       foregroundColor: Color(0xFFe5d6ca),
       backgroundGradient: [
         Color(0xFFbe916d),
         Color(0xFF785336),
       ],
     ),
-    AdwAvatarColors.gray: AdwAvatarColor(
+    AdwColors.gray: AdwAvatarColor(
       foregroundColor: Color(0xFFd8d7d3),
       backgroundGradient: [
         Color(0xFFc0bfbc),
@@ -152,11 +152,11 @@ class AdwAvatarColorPalette {
   /// Returns a set of values depending on the `color` variable.
   ///
   /// If the variable is `null`, it'll then return a random set of colors.
-  static AdwAvatarColor getColor(AdwAvatarColors? color) {
+  static AdwAvatarColor getColor(AdwColors? color) {
     if (color != null) return _palette[color]!;
 
     final randomColor =
-        AdwAvatarColors.values[Random().nextInt(AdwAvatarColors.values.length)];
+        AdwColors.values[Random().nextInt(AdwColors.values.length)];
 
     return _palette[randomColor]!;
   }
@@ -182,7 +182,7 @@ class AdwAvatar extends StatelessWidget {
     Key? key,
     required String text,
     double size = defaultAvatarSize,
-    AdwAvatarColors? backgroundColor,
+    AdwColors? backgroundColor,
     ImageProvider? backgroundImage,
   }) {
     assert(text.isNotEmpty, 'Text should not be empty');
@@ -220,7 +220,7 @@ class AdwAvatar extends StatelessWidget {
   /// Main color that will be used to decorate the widget.
   ///
   /// If `null`, a random set of colors will be picked.
-  final AdwAvatarColors? backgroundColor;
+  final AdwColors? backgroundColor;
 
   /// Image that will be rendered at the background of the widget.
   ///