+++ lib/src/widgets/adw/view_switcher.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:libadwaita/libadwaita.dart';
+import 'package:libadwaita/src/models/models.dart';
+import 'package:libadwaita/src/widgets/widgets.dart';
 
 class AdwViewSwitcher extends StatelessWidget {
   const AdwViewSwitcher({
@@ -7,6 +8,7 @@ class AdwViewSwitcher extends StatelessWidget {
     required this.tabs,
     required this.onViewChanged,
     required this.currentIndex,
+    this.badgeColor,
     this.style,
     @Deprecated('This parameter is no longer in use')
         double height = 55,
@@ -18,6 +20,7 @@ class AdwViewSwitcher extends StatelessWidget {
   })  : assert(tabs.length >= 2, 'Minimum 2 tabs are required'),
         super(key: key);
 
+  final Color? badgeColor;
   final List<ViewSwitcherData> tabs;
   final ValueChanged<int> onViewChanged;
   final ViewSwitcherStyle? style;
@@ -38,6 +41,7 @@ class AdwViewSwitcher extends StatelessWidget {
             for (final tab in tabs.asMap().entries)
               AdwViewSwitcherTab(
                 data: tab.value,
+                badgeColor: badgeColor,
                 style: newStyle,
                 isSelected: tab.key == currentIndex,
                 onSelected: currentIndex != tab.key