+++ lib/src/widgets/gtk/popup_menu.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:libadwaita/libadwaita.dart';
-import 'package:libadwaita/src/internal/popover.dart';
+import 'package:libadwaita/src/utils/colors.dart';
+import 'package:libadwaita/src/widgets/widgets.dart';
+import 'package:popover_gtk/popover_gtk.dart';
 
 class AdwPopupMenu extends StatefulWidget {
   const AdwPopupMenu({
@@ -39,11 +40,20 @@ class _AdwPopupMenuState extends State<AdwPopupMenu> {
         setState(() => isActive = true);
         showPopover(
           context: context,
-          child: widget.body,
+          shadow: [
+            BoxShadow(
+              color: context.borderColor,
+              blurRadius: 6,
+            ),
+          ],
+          barrierColor: Colors.transparent,
+          bodyBuilder: (_) => Padding(
+            padding: const EdgeInsets.all(4),
+            child: widget.body,
+          ),
           width: widget.popupWidth,
           height: widget.popupHeight,
           backgroundColor: Theme.of(context).cardColor,
-          contentOffset: const Offset(0, 4),
         ).whenComplete(() => setState(() => isActive = false));
       },
     );