+++ lib/src/widgets/widgets.dart
@@ -1,4 +1,7 @@
+// ignore_for_file: directives_ordering
+
 // All the widgets imported from libadwaita
+export 'adw/about_window.dart';
 export 'adw/action_row.dart';
 export 'adw/avatar.dart';
 export 'adw/button.dart';
@@ -6,17 +9,19 @@ export 'adw/clamp.dart';
 export 'adw/combo_row.dart';
 export 'adw/flap.dart';
 export 'adw/header_bar.dart';
+export 'adw/preferences_group.dart';
+export 'adw/switch.dart';
+export 'adw/view_stack.dart';
+export 'adw/view_switcher.dart';
+export 'adw/view_switcher_tab.dart';
+
+// New Widgets that are not present in libadwaita
 export 'adw/new/header_button.dart';
 export 'adw/new/scaffold.dart';
 export 'adw/new/sidebar.dart';
 export 'adw/new/text_button.dart';
-
-// New Widgets that are not present in libadwaita
 export 'adw/new/text_field.dart';
-export 'adw/preferences_group.dart';
-export 'adw/view_stack.dart';
-export 'adw/view_switcher.dart';
-export 'adw/view_switcher_tab.dart';
+export 'adw/new/window_button.dart';
 
 // All the widgets imported from Gtk4
 export 'gtk/popup_menu.dart';