+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## 1.0.0
+
+### **BREAKING**
+- `AdwHeaderBar.minimal` is now `AdwHeaderBar.custom`
+- Remove `label` parameter from `AdwTextField`
+
+### **Changes to widgets**
+**HeaderBar**
+- Now the `AdwHeaderBar` is not dependent on any package, `windowDecor` object is now optional
+- Add `isTransparent` parameter => Makes `AdwHeaderBar`'s background and border color
+
+**Popover**
+- Revisit popup menu by using `popover_gtk` package (popover package with fade transition) (#35)
+
+**ViewSwitcher**
+- Add `badge` in `AdwViewSwitcher`
+
+**New Widgets**
+- `AdwSwitch` => port of `GtkSwitch` from gtk4
+- `AdwAboutWindow` => port of upcoming `AdwAboutWindow` from libadwaita
+
 ## 1.0.0-rc.2
 
 **BREAKING**